requirements.txt
openai>=1.0.0
transformers>=4.40.0
torch>=2.2.0
datasets>=2.18.0
pyyaml>=6.0
tqdm>=4.66.0
accelerate>=0.29.0
